✨ (baseof.html): add data attributes for environment and version ♻️ (header.html): remove redundant environment output
Add data attributes `data-ndka-environment` and `data-ndka-version` to the `<body>` tag in `baseof.html` to provide easy access to the current Hugo environment and version information. This change enhances the ability to dynamically adjust the site's behavior or styling based on the environment or version.

Remove the redundant output of `hugo.Environment` in `header.html` as it is now included as a data attribute in the `<body>` tag, reducing clutter and potential confusion in the HTML structure.