✨ (site): add image resizing functionality and integrate with main menu
Introduce a new partial template for image resizing to optimize image
loading and improve performance. The image-resize.html partial allows
dynamic resizing of images based on specified dimensions, ensuring
images are served at the appropriate size. This change is integrated
into the main menu to replace the static image tag with the new
resizing functionality, enhancing the flexibility and maintainability
of the site. The addition of the nkdagility-with-martin-hinshelwood-light.png
image supports this new feature.