✨ (jsonld): add data-nkda-type attribute to script tag for enhanced metadata tracking 🖼️ (images): add new image nkdagility-with-martin-hinshelwood-dark-300px.png
The addition of the `data-nkda-type` attribute to the script tag in the JSON-LD template allows for better tracking and identification of the page type, which can be useful for analytics and debugging purposes. The new image file `nkdagility-with-martin-hinshelwood-dark-300px.png` is added to the static images directory, likely for use in the site's design or content.